+++ packages/lib/src/index.ts
@@ -1,9 +1,13 @@
-import { once } from 'lodash';
+import once from 'lodash/once';
 import * as typeIs from './typeIs';
 import createUniqueClassName from './createUniqueClassName';
 import getNodeTextContent from './getNodeTextContent';
+import DarkModeProps, { Theme } from './DarkModeProps';
+import getTheme from './getTheme';
 
-export { typeIs, createUniqueClassName, getNodeTextContent };
+export { typeIs, createUniqueClassName, getNodeTextContent, getTheme, Theme };
+export type { DarkModeProps };
+
 /** Helper type to extract an HTML element's valid props */
 export type HTMLElementProps<
   Element extends keyof JSX.IntrinsicElements,