+++ packages/tabs/src/Tabs.story.tsx
@@ -3,10 +3,11 @@ import { Meta } from '@storybook/react';
 import Icon from '@leafygreen-ui/icon';
 import { Tab, Tabs } from './index';
 import { TabsProps } from './Tabs';
+import defaultArgTypes from '../../../stories/defaultArgTypes';
 
 // TODO: Add subcomponent controls for Tab when supported by Storybook
 export default {
-  title: 'Packages/Tabs',
+  title: 'Components/Tabs',
   component: Tabs,
   args: {
     children: [
@@ -39,6 +40,7 @@ export default {
     as: { control: false },
     setSelected: { control: false },
     selected: { control: 'number' },
+    darkMode: defaultArgTypes.darkMode,
   },
   subcomponents: { Tab },
 } as Meta<typeof Tabs>;