+++ website/next.config.js
@@ -6,6 +6,7 @@ const withMDX = nextMdx({
   extension: /\.mdx?$/,
   options: {
     rehypePlugins: [rehypeSlug, [rehypeAutolinkHeadings, { behavior: 'wrap' }]],
+    providerImportSource: '@mdx-js/react',
   },
 });