+++ packages/combobox/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @leafygreen-ui/combobox
 
+## 2.0.1
+
+### Patch Changes
+
+- 5f82468e: Updates lodash imports in combobox. Lodash functions are now imported individually to reduce bundle size
+- Updated dependencies [13a4adcc]
+  - @leafygreen-ui/typography@13.0.1
+
+## 2.0.0
+
+### Major Changes
+
+- a0beace8: Updates Combobox for visual refresh. Adds Dark Mode
+
+### Patch Changes
+
+- Updated dependencies [85d46871]
+- Updated dependencies [99e20bb9]
+  - @leafygreen-ui/lib@9.4.0
+  - @leafygreen-ui/checkbox@9.0.0
+  - @leafygreen-ui/icon-button@13.0.0
+  - @leafygreen-ui/popover@9.0.0
+  - @leafygreen-ui/typography@13.0.0
+  - @leafygreen-ui/tooltip@7.0.4
+  - @leafygreen-ui/inline-definition@4.0.1
+
 ## 1.2.2
 
 ### Patch Changes