+++ packages/interaction-ring/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @leafygreen-ui/interaction-ring
 
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies [85d46871]
+- Updated dependencies [99e20bb9]
+  - @leafygreen-ui/lib@9.4.0
+  - @leafygreen-ui/leafygreen-provider@2.3.0
+
 ## 3.0.0
 
 ### Major Changes