+++ packages/radio-box-group/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @leafygreen-ui/radio-box-group
 
+## 9.0.1
+
+### Patch Changes
+
+- 1d471e2c: Removes stacking context from Radio Box
+
+## 9.0.0
+
+### Patch Changes
+
+- Updated dependencies [85d46871]
+- Updated dependencies [99e20bb9]
+  - @leafygreen-ui/lib@9.4.0
+  - @leafygreen-ui/leafygreen-provider@2.3.0
+
 ## 8.0.2
 
 ### Patch Changes